+++ babel.config.js
@@ -9,6 +9,7 @@ module.exports = function(api) {
           'root': ['./'], 
           'alias': {
             '@src': './src',
+            '@utils': './src/utils',
             '@assets': './assets',
             '@data': './src/data',
             '@components': './src/components'
