+++ src/components/MarkerComponent.js
@@ -2,8 +2,6 @@ import React from 'react';
 import { StyleSheet, View, Text } from 'react-native';
 import { Marker, Callout } from 'react-native-maps';
 
-import placeInformation from '@data/placesOfInterest';
-
 const defaultDescription = "need to add a description for this awesome place!";
 
 export default function(props) {
