--- src/data/mapData.js
@@ -1,29 +1,0 @@
-// export default KML_DATA = {
-//   "./assets/kml/TNQ1alto2ha.kml": require("./assets/kml/TNQ1alto2ha.kml"),
-//   "./assets/kml/GLperimetro.kml": require("./assets/kml/GLperimetro.kml"),
-//   "./assets/kml/PTOAGUA.kml": require("./assets/kml/PTOAGUA.kml")
-// }
-
-export const mapOverlayCoordinates = [
-  { 
-    longitude: -77.89770174815041,
-    latitude: -6.054473497588077 + .00001 // adjustar un poco
-  },
-  // { 
-  //   longitude: -77.89744399531243,
-  //   latitude: -6.054210153938867
-  // },
-  // { 
-  //   longitude: -77.89790378080357,
-  //   latitude: -6.055070647890809
-  // },
-  { 
-    longitude: -77.89645032241788,
-    latitude: -6.055485813090427
-  }
-];
-
-//RT -77.89771312572236,-6.05546647110471,0 
-// RB -77.89645032241788,-6.055485813090427,0 
-//LB -77.89642592047115,-6.05447009687464,0 
-//LT -77.89770174815041,-6.054473497588077,0
