--- src/data/placesOfInterest.js
@@ -1,7 +1,0 @@
-export default {
-  0: "Celery quandong swiss chard chicory earthnut pea potato. Salsify taro catsear garlic gram celery bitterleaf wattle seed collard greens nori.\n Grape wattle seed kombu beetroot horseradish carrot squash brussels sprout chard.",
-  1: "un descripcion de algo",
-  2: "arbol con buenas propeidades y usado como cura de malaria",
-  3: "arbol con buenas propeidades y usado como cura de malaria",
-  4: "This garden was made by one of our residents, a Japanese gardener named ...\n Here you will find a variety of succulent species - Gumbo beet greens corn soko endive gumbo gourd. Parsley shallot courgette tatsoi pea sprouts fava bean collard greens dandelion okra wakame tomato. Dandelion cucumber earthnut pea peanut soko zucchini."
-}
